+++ types/browan/TBDW100/uplink.js
@@ -1,3 +1,17 @@
+function calculateIncrement(lastValue, currentValue) {
+  // Check if current value exists
+  if (currentValue === undefined || Number.isNaN(currentValue)) {
+    return 0;
+  }
+
+  // Init state && Check for the case the counter reseted
+  if (lastValue === undefined || lastValue > currentValue) {
+    lastValue = currentValue;
+  }
+  // Calculate increment
+  return currentValue - lastValue;
+}
+
 function consume(event) {
   const payload = event.data.payloadHex;
   const bits = Bits.hexToBits(payload);
@@ -21,10 +35,14 @@ function consume(event) {
 
   data.temperature = Bits.bitsToUnsigned(bits.substr(17, 7));
   data.temperature -= 32;
-
   data.time = Hex.hexLittleEndianToBigEndian(payload.substr(6, 4), false);
   data.count = Hex.hexLittleEndianToBigEndian(payload.substr(10, 6), false);
 
+  const state = event.state || {};
+  data.relativeCount = calculateIncrement(state.lastCount, data.count);
+  state.lastCount = data.count;
+
+  emit("state", state);
   emit("sample", { data: lifecycle, topic: "lifecycle" });
   emit("sample", { data, topic: "default" });
 }