+++ types/decentlab/DWS/uplink.spec.js
@@ -53,9 +53,9 @@ describe("Decentlab DWS Uplink", () => {
         assert.equal(value.topic, "default");
         assert.equal(value.data.frequency, 15350.47);
         assert.equal(value.data.weight, 11.26);
-        assert.equal(value.data.weightKg, 0.01126);
-        assert.equal(value.data.incrementGram, 0);
-        assert.equal(value.data.incrementKg, 0);
+        assert.equal(value.data.weightKg, 0.0113);
+        assert.equal(value.data.relativeWeightGram, 0);
+        assert.equal(value.data.relativeWeightKilogramm, 0);
 
         utils.validateSchema(value.data, defaultSchema, { throwError: true });
       });
@@ -106,9 +106,9 @@ describe("Decentlab DWS Uplink", () => {
         assert.equal(value.topic, "default");
         assert.equal(value.data.frequency, 15350.47);
         assert.equal(value.data.weight, 11.26);
-        assert.equal(value.data.weightKg, 0.01126);
-        assert.equal(value.data.incrementGram, 6.26);
-        assert.equal(value.data.incrementKg, 0.006259999999999999);
+        assert.equal(value.data.weightKg, 0.0113);
+        assert.equal(value.data.relativeWeightGram, 6.26);
+        assert.equal(value.data.relativeWeightKilogramm, 0.0063);
 
         utils.validateSchema(value.data, defaultSchema, { throwError: true });
       });