+++ types/netvox/R718N3/uplink.spec.js
@@ -5,23 +5,17 @@ const utils = require("test-utils");
 const { assert } = chai;
 
 describe("Netvox R718N3 Uplink", () => {
-  let raw1Schema = null;
+  let defaultSchema = null;
   let consume = null;
   before((done) => {
     const script = rewire("./uplink.js");
     consume = utils.init(script);
-    utils.loadSchema(`${__dirname}/raw1.schema.json`).then((parsedSchema) => {
-      raw1Schema = parsedSchema;
-      done();
-    });
-  });
-
-  let raw2Schema = null;
-  before((done) => {
-    utils.loadSchema(`${__dirname}/raw2.schema.json`).then((parsedSchema) => {
-      raw2Schema = parsedSchema;
-      done();
-    });
+    utils
+      .loadSchema(`${__dirname}/default.schema.json`)
+      .then((parsedSchema) => {
+        defaultSchema = parsedSchema;
+        done();
+      });
   });
 
   let lifecycleSchema = null;
@@ -30,16 +24,6 @@ describe("Netvox R718N3 Uplink", () => {
       .loadSchema(`${__dirname}/lifecycle.schema.json`)
       .then((parsedSchema) => {
         lifecycleSchema = parsedSchema;
-        done();
-      });
-  });
-
-  let defaultSchema = null;
-  before((done) => {
-    utils
-      .loadSchema(`${__dirname}/default.schema.json`)
-      .then((parsedSchema) => {
-        defaultSchema = parsedSchema;
         done();
       });
   });
@@ -57,6 +41,7 @@ describe("Netvox R718N3 Uplink", () => {
   describe("consume()", () => {
     it("should decode the first raw R718N3 payload", () => {
       const data = {
+        state: {},
         data: {
           port: 1,
           payloadHex: "014a01243f803fb03f8001",
@@ -64,17 +49,12 @@ describe("Netvox R718N3 Uplink", () => {
       };
 
       utils.expectEmits((type, value) => {
-        assert.equal(type, "sample");
-        assert.isNotNull(value);
-        assert.typeOf(value.data, "object");
-        assert.equal(value.topic, "raw1");
+        assert.equal(type, "state");
 
-        assert.equal(value.data.current1, 16256);
-        assert.equal(value.data.current2, 16304);
-        assert.equal(value.data.current3, 16256);
-        assert.equal(value.data.multiplier1, 1);
-
-        utils.validateSchema(value.data, raw1Schema, { throwError: true });
+        assert.equal(value.current1, 16256);
+        assert.equal(value.current2, 16304);
+        assert.equal(value.current3, 16256);
+        assert.equal(value.multiplier1, 1);
       });
 
       utils.expectEmits((type, value) => {
@@ -93,6 +73,12 @@ describe("Netvox R718N3 Uplink", () => {
 
     it("should decode the second raw R718N3 payload", () => {
       const data = {
+        state: {
+          current1: 16256,
+          current2: 16304,
+          current3: 16256,
+          multiplier1: 1,
+        },
         data: {
           port: 1,
           payloadHex: "014a022401010000000000",
@@ -103,12 +89,13 @@ describe("Netvox R718N3 Uplink", () => {
         assert.equal(type, "sample");
         assert.isNotNull(value);
         assert.typeOf(value.data, "object");
-        assert.equal(value.topic, "raw2");
+        assert.equal(value.topic, "default");
 
-        assert.equal(value.data.multiplier2, 1);
-        assert.equal(value.data.multiplier3, 1);
+        assert.equal(value.data.current1, 16.256);
+        assert.equal(value.data.current2, 16.304);
+        assert.equal(value.data.current3, 16.256);
 
-        utils.validateSchema(value.data, raw2Schema, { throwError: true });
+        utils.validateSchema(value.data, defaultSchema, { throwError: true });
       });
 
       utils.expectEmits((type, value) => {