+++ types/swisscom/multisense/uplink.js
@@ -7,12 +7,27 @@ function int16(hex) {
   return a;
 }
 
+function calculateIncrement(lastValue, currentValue) {
+  // Check if current value exists
+  if (currentValue === undefined || Number.isNaN(currentValue)) {
+    return 0;
+  }
+
+  // Init state && Check for the case the counter reseted
+  if (lastValue === undefined || lastValue > currentValue) {
+    lastValue = currentValue;
+  }
+  // Calculate increment
+  return currentValue - lastValue;
+}
+
 function consume(event) {
   const payload = event.data.payloadHex;
   const { port } = event.data;
   const bits = Bits.hexToBits(payload);
   const lifecycle = {};
   const trigger = {};
+  const state = event.state || {};
   let topic = "default";
 
   lifecycle.payloadVersion = Bits.bitsToUnsigned(bits.substr(0, 8));
@@ -58,6 +73,13 @@ function consume(event) {
           data.reedCounter = Bits.bitsToUnsigned(bits.substr(pointer, 16));
           pointer += 16;
           topic = "reed_counter";
+
+          data.relativeReedCounter = calculateIncrement(
+            state.lastReed,
+            data.reedCounter,
+          );
+          state.lastReed = data.reedCounter;
+
           break;
         case 4:
           data.motionCounter = Bits.bitsToUnsigned(bits.substr(pointer, 16));
@@ -152,4 +174,5 @@ function consume(event) {
 
   emit("sample", { data: trigger, topic: "event" });
   emit("sample", { data: lifecycle, topic: "lifecycle" });
+  emit("state", state);
 }